@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "React & TypeScript application starter",
   "scripts": {
-    "start": "webpack-dev-server --open",
+    "start": "webpack serve --open",
     "build": "NODE_ENV='production' ./node_modules/.bin/webpack --config webpack.config.js",
     "build-dev": "NODE_ENV='development' ./node_modules/.bin/webpack --config webpack.config.js",
     "test": "npm run test-only",
@@ -13,41 +13,44 @@
   "devDependencies": {
     "@babel/core": "^7.9.0",
     "@babel/preset-env": "^7.9.0",
+    "@babel/preset-react": "^7.12.13",
+    "@babel/preset-typescript": "^7.13.0",
     "@types/enzyme": "^3.9.0",
     "@types/enzyme-adapter-react-16": "^1.0.5",
-    "@types/jest": "^25.1.4",
-    "ajv": "^6.6.1",
-    "autoprefixer": "^9.7.5",
+    "@types/jest": "^26.0.20",
+    "@types/react": "16.9.19",
+    "@types/react-dom": "16.9.5",
+    "ajv": "^7.1.1",
     "babel-loader": "^8.1.0",
-    "css-loader": "^3.4.2",
+    "css-loader": "^5.0.2",
     "enzyme": "^3.4.1",
     "enzyme-adapter-react-16": "^1.10.0",
     "express": "^4.16.3",
-    "html-webpack-plugin": "^4.0.1",
+    "html-webpack-plugin": "^5.2.0",
+    "import-sort-cli": "6.0.0",
     "import-sort-parser-typescript": "^6.0.0",
     "import-sort-style-as": "^1.2.3",
-    "jest": "^25.1.0",
+    "jest": "^26.6.3",
     "jest-enzyme": "^7.0.2",
-    "mini-css-extract-plugin": "^0.9.0",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
+    "mini-css-extract-plugin": "^1.3.8",
+    "postcss": "^8.2.6",
+    "postcss-import": "^14.0.0",
+    "postcss-loader": "^5.0.0",
     "postcss-preset-env": "^6.6.0",
     "precss": "^4.0.0",
-    "style-loader": "^1.1.3",
-    "ts-jest": "^25.2.1",
-    "ts-loader": "^6.2.2",
-    "typescript": "^3.8.3",
+    "style-loader": "^2.0.0",
+    "ts-jest": "^26.5.2",
+    "ts-loader": "^8.0.17",
+    "typescript": "^4.2.2",
     "url-loader": "^4.0.0",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.10.3",
-    "webpack-merge": "^4.1.3"
+    "webpack": "^5.24.1",
+    "webpack-cli": "^4.5.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.7.3"
   },
   "dependencies": {
-    "@types/react": "^16.9.27",
-    "@types/react-dom": "^16.9.5",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react": "16.14.0",
+    "react-dom": "16.14.0"
   },
   "prettier": {
     "trailingComma": "es5",
